Add cuSPARSE coloring implementation
Change verteces type from size_t to int

+++ ColoringAlgorithm/ColoringAlgorithm.h
@@ -24,9 +24,9 @@ public:
 
 	// Receive vertex index, number of colors used so far and pointer where to save the color.
 	// Returns the number of colors used
-	const int computeVertexColor(size_t const v, int const n_cols, int* targetCol) const;
+	const int computeVertexColor(int const v, int const n_cols, int* targetCol) const;
 
-	std::vector<std::pair<size_t, size_t>> checkCorrectColoring();
+	std::vector<std::pair<int, int>> checkCorrectColoring();
 
 	void printColors(std::ostream& os) const;
 	void printDotFile(std::ostream& os) const;
