+++ ColoringAlgorithm/GebremedhinManne.cpp
@@ -123,7 +123,7 @@ int GebremedhinManne::detectConflicts() {
 }
 
 void GebremedhinManne::detectConflictsParallel(const int i) {
-	for (size_t v = i; v < this->adj().nV(); v += this->MAX_THREADS_SOLVE) {
+	for (int v = i; v < this->adj().nV(); v += this->MAX_THREADS_SOLVE) {
 		if (this->col[v] == GebremedhinManne::INVALID_COLOR) {
 			this->mutex.lock();
 			this->recolor.push_back(v);
@@ -134,7 +134,7 @@ void GebremedhinManne::detectConflictsParallel(const int i) {
 		auto const end = this->adj().endNeighs(v);
 		for (auto neighIt = this->adj().beginNeighs(v);
 			neighIt != end;	++neighIt) {
-			size_t w = *neighIt;
+			int w = *neighIt;
 			if (v < w) continue;
 
 			if (this->col[v] == this->col[w]) {
