+++ GraphRepresentation/CompressedSparseRow.h
@@ -12,35 +12,35 @@
 
 class CompressedSparseRow : public GraphRepresentation {
 private:
-	size_t rows;
-	size_t cols;
+	int rows;
+	int cols;
 
-	std::vector<size_t> col_idxs;
-	size_t* row_ptrs;
+	std::vector<int> col_idxs;
+	int* row_ptrs;
 
-	void init(size_t n_rows, size_t n_cols);
+	void init(int n_rows, int n_cols);
 
 	template<typename Iterator>
-	void populateRow(size_t row_idx, Iterator begin, const Iterator end);
+	void populateRow(int row_idx, Iterator begin, const Iterator end);
 
 public:
 	CompressedSparseRow();
 	CompressedSparseRow(const CompressedSparseRow& to_copy);
-	CompressedSparseRow(size_t rows, size_t cols);
-	CompressedSparseRow(size_t n);
+	CompressedSparseRow(int rows, int cols);
+	CompressedSparseRow(int n);
 	~CompressedSparseRow();
 
-	const size_t* getColIndexes() const;
-	const size_t* getRowPointers() const;
+	const int* getColIndexes() const;
+	const int* getRowPointers() const;
 
 	friend std::istream& operator>>(std::istream& is, CompressedSparseRow& m);
 
 	const size_t nE() const override;
-	const size_t nV() const override;
-	const bool get(size_t row, size_t col) const override;
+	const int nV() const override;
+	const bool get(int row, int col) const override;
 
-	const ::std::vector<size_t>::const_iterator beginNeighs(size_t row_idx) const override;
-	const ::std::vector<size_t>::const_iterator endNeighs(size_t row_idx) const override;
+	const ::std::vector<int>::const_iterator beginNeighs(int row_idx) const override;
+	const ::std::vector<int>::const_iterator endNeighs(int row_idx) const override;
 };
 
 #endif // !_COMPRESSED_SPARSE_ROW_H
