+++ GraphRepresentation/GraphRepresentation.h
@@ -15,17 +15,17 @@ public:
 	// Returns total number of edges
 	virtual const size_t nE() const = 0;
 	// Returns total number of verteces
-	virtual const size_t nV() const = 0;
+	virtual const int nV() const = 0;
 	// Returns whether v is neighbor of w
-	virtual const bool get(size_t v, size_t w) const = 0;
+	virtual const bool get(int v, int w) const = 0;
 
-	const int countNeighs(size_t v) const;
+	const int countNeighs(int v) const;
 	static void printGraphRepresentationConfs();
 
 	// Returns begin iterator of neighbors of v
-	virtual const ::std::vector<size_t>::const_iterator beginNeighs(size_t v) const = 0;
+	virtual const ::std::vector<int>::const_iterator beginNeighs(int v) const = 0;
 	// Returns end iterator of neighbors of v
-	virtual const ::std::vector<size_t>::const_iterator endNeighs(size_t v) const = 0;
+	virtual const ::std::vector<int>::const_iterator endNeighs(int v) const = 0;
 };
 
 #endif // !_GRAPH_REPRESENTATION_H
